refactor AjaxFileUpload tests
