+++ AjaxControlToolkit.Tests/AjaxControlToolkit.Tests.csproj
@@ -77,7 +77,9 @@
     <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AjaxFileUploadTests.cs" />
+    <Compile Include="AjaxFileUpload\AjaxFileUploadTests.cs" />
+    <Compile Include="AjaxFileUpload\AjaxFileUploadWrapper.cs" />
+    <Compile Include="AjaxFileUpload\AppDomainBuilder.cs" />
     <Compile Include="BundleResolverTests.cs" />
     <Compile Include="BaseValidatorTests.cs" />
     <Compile Include="ComponentDescriber\ComponentDescriberTests.cs" />
