+++ AjaxControlToolkit.Tests/AjaxFileUpload/AjaxFileUploadWrapper.cs
@@ -1,0 +1,42 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+
+namespace AjaxControlToolkit.Tests {
+
+    class AjaxFileUploadWrapper : MarshalByRefObject {
+
+        public void ProcessStreamWithoutTempRootPath() {
+            AjaxFileUploadHelper.RootTempFolderPath = "";
+            var stream = GenerateStreamFromString("------WebKitFormBoundaryuzPlX1oHHDDbSusw\r\nContent-Disposition: form-data; name=\"act-file-data\"; filename=\"1#.txt\"\r\nContent-Type: text/plain\r\n\r\n123\r\n------WebKitFormBoundaryuzPlX1oHHDDbSusw--\r\n");
+            new AjaxFileUploadHelper().ProcessStream(new FakeCache(), stream, "fileId", "fileName", false, false, false);
+        }
+
+        public void ProcessStreamWithTempRootPath() {
+            AjaxFileUploadHelper.RootTempFolderPath = @"C:\";
+            var stream = GenerateStreamFromString("------WebKitFormBoundaryuzPlX1oHHDDbSusw\r\nContent-Disposition: form-data; name=\"act-file-data\"; filename=\"1#.txt\"\r\nContent-Type: text/plain\r\n\r\n123\r\n------WebKitFormBoundaryuzPlX1oHHDDbSusw--\r\n");
+            new AjaxFileUploadHelper().ProcessStream(new FakeCache(), stream, "fileId", "fileName", false, false, false);
+        }
+
+        MemoryStream GenerateStreamFromString(string value) {
+            return new MemoryStream(Encoding.UTF8.GetBytes(value ?? ""));
+        }
+
+        private class FakeCache : IWebCache {
+
+            Dictionary<string, object> set = new Dictionary<string, object>();
+
+            public object this[string key] {
+                get {
+                    if(set.ContainsKey(key))
+                        return set[key];
+
+                    return null;
+                }
+                set { set[key] = value; }
+            }
+        }
+    }
+}
